requirement.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
